src/event/requirements.txt
pymysql
google-api-python-client 
google-auth-httplib2 
google-auth-oauthlib
pytz